Fix a rather nastly bug in REPL environment saving
Popping up an error from a different thread was misteriously updaing a
REPL's environment (*, **, etc...) but only when the debugger was quit.

Figured out that WITH-DEFAULT-LISTENER called WITH-LISTENER which
updates bindings of the listener at the end, even in situations where
we just want to read them.

The new system tries to clearly identify which places we want a
listener's environment to change.

* contrib/slynk-mrepl.lisp (mrepl-eval-1): Update comment.
(globally-redirect-to-listener, with-eval-for-repl, create-mrepl)
(clear-repl-history)
(mrepl-eval): Use SAVING-LISTENER-BINDINGS.
(:inspect-object, with-eval-for-repl, maybe-redirect-global-io):
Use WITH-LISTENER-BINDINGS.

* slynk/slynk.lisp (with-default-listener): Use
WITH-LISTENER-BINDINGS.
(:compile-toplevel): Export WITH-LISTENER-BINDINGS and
SAVING-LISTENER-BINDINGS.
(with-listener-bindings, saving-listener-bindings)
(call-with-listener): New macros and helpers.
(with-listener): Deleted.