Introduce break-on-stickers
Use with `M-x sly-stickers-toggle-break-on-stickers'. Yet to be
improved: currently the sticker that pops the debugger isn't focused
into view, neither can its values be inspected or returned to the
REPL.

* contrib/sly-stickers.el (sly-stickers-replay): Fix another bug
when returning sticker recordings to REPL.
(sly-stickers-toggle-break-on-stickers): New interactive function.

* contrib/slynk-stickers.lisp (sticker): Add IGNORE-SPEC slot.
(print-object): Nicer print function maybe.
(compile-for-stickers): Improve docstring.
(just-before-sticker, right-after-sticker): New conditions.
(*break-on-stickers*): New parameter.
(toggle-break-on-stickers): New slyfun.
(call-with-sticker-recording): Honour *BREAK-ON-STICKERS* and
their IGNORE-SPEC slots.
(record): Use ~S, not ~A