Closes #64: Fix regular isearch in the REPL
The hack for working around emacs bug 19572 is very aggresive (and
probably not needed after 24.5), and as such was preventing correct
isearch in the REPL. Fixed that. The hack should be removed pretty
soon.

* sly-mrepl.el (sly-mrepl--setup-comint-isearch): Don't
work around emacs bug 19572 when there isn't a need to.
(sly-mrepl--teardown-comint-isearch): Reset the hacks done in
`sly-mrepl--setup-comint-isearch'.