Closes #110: Don't let paredit.el break REPL restarts
When `sly-restart-inferior-lisp' is called, a new REPL is launched
that potentially reuses an old REPL buffer, potentially full of
unmatched parens.

Well, paredit.el doens't like this and, if re-launched in those
conditions, will check parens and trigger an error that chokes the
lisp restart.

We could try to prevent that by not needlessly re-invoking
`sly-mrepl-mode' for a buffer already in that mode, but that currently
breaks some tests for some unknown reason. So temporarily set
paredit's variable `paredit-override-check-parens-function' to exempt
`sly-mrepl-mode' from the problematic checks.

See issue #110 for more insights.

* contrib/sly-mrepl.el(sly-mrepl-mode): Set
paredit-override-check-parens-function to not error when
unbalances are discovered on REPL restart.