Parameters controlling REPL output can be set from ~/.slynkrc
Noticed this while debugging #113. The solution to change
*DEDICATED-OUTPUT-STREAM-BUFFERING* couldn't easily be realized if the
symbols live in SLYNK-MREPL, which isn't known at the time the user
init file is read in.

* doc/sly.texi (REPL output control): Parameters controlling REPL
output are (exceptionally) symbols in the SLYNK package, for easy
setting from the user's init file.
(Setting up the Lisp image): Correct package designator for
SLYNK:*DEDICATED-OUTPUT-STREAM-PORT*

* slynk/slynk.lisp (:slynk): Export (but don't define)
*DEDICATED-OUTPUT-STREAM-PORT*,
*DEDICATED-OUTPUT-STREAM-BUFFERING*,
*USE-DEDICATED-OUTPUT-STREAM*.