Closes #123: M-. jumps more closely respect window and frame intents
Previous to this commit `sly-edit-defition' would sometime use the
same window to display a location or a different one, according to the
wheter an xref buffer would result of a definition search or not.

Since the different commands `sly-edit-definition',
`sly-edit-definition-other-window' and
`sly-edit-definition-other-frame' already give the user a choice, make
this behaviour more consistent: if just `sly-edit-definition' was
issued, attempt to open in the original window (if it is still
available). Else try *not* to open in the original window.

This also supersedes the previous commit fixing #121, which was a temp
fix

* sly.el (sly--maybe-reuse-maybe-popup-window): Delete.
(sly-with-popup-buffer): Fix bug and rewrite a bit.
(sly--xref-popup-method): New local variable.
(sly--pop-to-source-location): Rewrite
(sly-edit-definition): Pass METHOD to sly-xref--show-results
(sly-edit-uses): Pass 'window to sly-edit-uses
(sly-xref-goto, define-button-type sly-xref): Pass 'sly--xref to
sly--pop-to-source-location.
(sly-xref--show-results): Set sly--xref-popup-method
(sly-xref-last-buffer): Delete

* lib/sly-tests.el (sly-recompile-all-xrefs): Fix test after
deleting sly-xref-last-buffer