Hopefully fix unstable indentation tests for a long time
The byte-compiler-warning fix of sly-cl-indent.el, in commit
227dc895b4fa4a23a3184185d5f42cb2282f3a3f, replaced all instances of
'containing-sexp' with sly-cl-indent--containing-sexp, which is
correct in all but one, which contained a bug. Luís Oliveira's SLIME
commit 8f19c4be949c161228453a533f3d6fdaed2a4523 corrects it.

Once this is corrected, the hacks of SLY commit
3e860fc244e6577b5da447237ef8568442e6323a are no longer needed, and so
are reverted.

The only outlier is test indent-60, which needs to be run with
indent-region, and indent-sexp will not work.

SLIME runs all these tests with indent-region, SLY works correctly
with indent-sexp or indent-region.

* contrib/sly-cl-indent.el (common-lisp-indent-function-1): Use
`containing-form-start' not `sly-cl-indent--containing-sexp'. This
is SLIME's commit 8f19c4be949c161228453a533f3d6fdaed2a4523.

* test/sly-cl-indent-test.txt (indent-60) needs to be done with
indent-region. (indent-84, indent-85): Restore correct expectation.

* test/sly-indentation-tests.el (sly-indentation-test--1): New function.
(sly-indentation-test-form): Expand to call to it.