Close #134: Unbreak newest ASDF that complains on recursive loads
The SLYNK-REQUIRE call in some contribs (actually only
contrib/slynk-fancy-inspector.lisp) is the pre-ASDF way to specify
dependencies between Slynk components, but causes a recursive ASDF
load (where the dependencies are in .asd files). For some reason the
brand new ASDF doesn't like this, so use a special variable to work
around it and just nop on the recursive load.

Once non-ASDF loading is removed in the near future, this won't be
necessary any more.

* slynk/slynk.lisp (*asdf-load-in-progress*): New variable.
(require-module :asdf): Use it.