Closes #135: Reloading Slynk on SBCL messes up threads
Loading some systems on SBCL 1.4.2 somehow forces the slynk-backend
ASDF system to be reloaded, and thus sbcl.lisp to be reloaded. Due to
a spurious DEFPARAMETER in this file, this reset an important
variable, SLYNK-SBCL::*THREAD-ID-MAP*, thus making most thread lookups
fail, leading to the behaviour described in the issue.

* slynk/backend/sbcl.lisp (*thread-id-map*): Make it a defvar.