Close #135: Reloading Slynk on SBCL messes up threads
For reasons still under investigation, the very newest ASDF in SBCL
1.4.2 forces the :SLYNK ASDF system to be needlessly reloaded (it is
already loaded by M-x sly using ASDF:LOAD-SYSTEM).

Regardless, this causes thus sbcl.lisp to be reloaded and this resets
an important variable, SLYNK-SBCL::*THREAD-ID-MAP*, which was
incorrectly defined with DEFPARAMETER. The effect is that most thread
lookups using ids before the reload fail, and this leads to the
behaviour described in the issue.

A good first solution is to use DEFVAR instead of DEFPARAMETER in
sbcl.lisp, then understand why ASDF is needlessly reloading :SLYNK.

* slynk/backend/sbcl.lisp (*thread-id-map*): Make it a defvar.