Disable part of the fix to #123 which is problematic in Emacs 26.1
The fix to emacs' bug#28814 in Emacs 26.1, which I myself authored,
causes problems with some approaches to protecting the original window
in the case of sly-edit-definition-other-window.

Best to disable it for now and eventually copy Emacs' approach to xref
buffers, or better yet, reuse it.

* sly.el (sly--display-source-location): Don't protect the
original window from reuse, it's not needed most of the time.