Fix Lisp mode comment indentation in newer emacs
There was a bug, introduced in emacs 26, that caused a single-semicolon comment
to break lisp-mode; it would cause the cursor to move to the start of the line
for each character inserted. It turns out that it can be fixed by removing an
obsolete workaround in slime-cl-indent's "basic" common-lisp-style.

(cherry picked from SLIME commit
17b88548e4db50c99597d616940aebf028a60dae)

* lib/sly-cl-indent.el: do the above things.