Fix Hash Table Inspector with Complex Numbers as keys
The EMACS-INSPECT method that specialized on HASH-TABLE tries to order
the contents of the hash-table before to present them in order. However
it incorrectly applies the function < if the contents are NUMBERs. Only =
and /= are defined for NUMBERS. < and the other comparison functions
apply only to REALs. Because this feature is an extra nicety fix the bug
by not ordering the contents in the presences of COMPLEX numbers.

To reproduce the bug inspect the hash table from return by

(let ((ht (make-hash-table)))
  (setf (gethash #C(0 1) ht) t
            (gethash #C(0 2) ht) t)
  ht)