Throttle function signature requests when using company
By default company has company-idle-delay set to 0.5 which is a
reasonable value.  It has no such throttling in place for candidate
signature requests (:company-docsig in company-capf), so if the user
holds down C-n (sometimes bound to company-select-next) it will make
many useless function signature requests to the server.

This commit adds a kind of very basic throttling capability to SLY, by
making some sly--responsive-eval requests sit for a small amount of
time before actually sending the request.  If user input arrives
within that time, we assume the user is not interested in the result
and return some quick value (usually nil).

* sly.el (sly-eval): Accept new parameter CANCEL-ON-INPUT-SIT-FOR.

* lib/sly-completion.el (sly--responsive-eval): Accept parameter
SIT-FOR.
(sly--completions-complete-symbol-1): Pass SIT-FOR to
sly--responsive-eval in :company-docsig.