Work around another Allegro condition location bug
When compiling a file via C-c C-k that contains only this:

   (defmethod foo (bla)
     (declare (ignore bla))
     (let (another-unused)
       (when nil)
       ))

Allegro would wrongly place the condition about ANOTHER-UNUSED being
unused, sometimes outside the file's boundaries.  Whereas this makes
it place it semi-correctly on the LET.

   (defmethod foo (bla)
     (declare (ignore bla))
     (let (another-unused)))

The workaround added by this change places it on the FOO, since this
bit of info isn't lost.

* slynk/backend/allegro.lisp (condition-pathname-and-position):
Work around an Allegro 10.1 bug.