Close #131: paper over SLY version detection issues when using MELPA
When installing via MELPA (but only in some situations that couldn't
be precisely determined) sly-protocol-version becomes nil and annoys
users with a useless message: "[sly] Versions differ: nil (sly) .."

Since this is most commonly a false warning which affects first time
users who may not care to look for and customize
sly-ignore-protocol-mismatches, we decide paper over this check when
sly-protocol-version is nil.

* sly.el (sly-check-version): Skip version check
if sly-protocol-version is nil.