Fix #211, #241: Review SLYNK-PPRINT logic and limits
SLYNK:ECHO-FOR-EMACS was needlessly calling
SLYNK:SLYNK-PPRINT-TO-LINE.  I don't think that's what the user wants.
Generally simplified limits, giving control to the user via
SLYNK:*BACKTRACE-PRINTER-BINDINGS* and SLYNK:*SLYNK-PPRINT-BINDINGS*.

Leave the protection in SLYNK:SLYNK-PPRINT and enhance it: if
*PRINT-CIRCLE* is nil, use (reasonably generous) limits.

* slynk/slynk.lisp (slynk-pprint): Don't wastefully, recursively
call slynk-pprint.  Better failsafe.
(echo-for-emacs): Accept FN, defaults to SLYNK-PPRINT.
(slynk-pprint-to-line): Don't mess with *PRINT-RIGHT-MARGIN*.
(frame-locals-for-emacs): Property use SLYNK-PPRINT.