Prevent some Elisp string properties from reaching Lisp
...where they completely break things.  Filenames are an example of
strings handed to Lisp that might have such properties: sly-tramp used
in conjuntion with Helm seems to provoke this.

The best way would be to do it at a lower level, in
sly-prin1-to-string, but it doesn't seem possible for now, so just
unpropertize strings in sly-to-lisp-filename.

Per #257 and #256.  Thanks to Alexander Artemenko for the original
implementation.

* sly.el (sly-to-lisp-filename): Use substring-no-properties.