Fix race condition with Allegro SMP
The problem is created when write to MP:PROCESS-PROPERTY-LIST isn't
properly serialized.

Specifically, even though SLYNK-ALLEGRO:MAILBOX was serialized with a
lock, this could still lead two threads calling SLYNK-ALLEGRO:MAILBOX
to create two different mailboxes, intead of the desired behaviour
where one just getting the one create by the other.  This could happen
with three calls in this order

1) SLYNK-ALLEGRO:MAILBOX creates a mailbox and attempts to push it
onto the plist, and returns it to Slynk for sending a message.

2) In the meantime SLYNK-ALLEGRO:THREAD-ID creates an ID and attempts
to push it onto the plist, overriting the efforts of (1).

3) A second SLYNK-ALLEGRO:MAILBOX decides to create a mailbox,
attempts to push it onto the plist, succeeds and returns it to Slynk
for waiting on messages.

The problem now is that the thread that did (3) is now waiting on a
different mailbox with SLYNK-BACKEND:RECEIVE-IF for a message that
will never appear there, because it was sent to the mailbox created in (1).

The fix is to use a global lock for the MP:PROCESS-PROPERTY-LIST.

* slynk/backend/allegro.lisp (*process-plist-lock*): New lock.
(mailbox, thread-id): Use it.