Fix race condition with Allegro SMP
The problem is created because reading and writing to
MP:PROCESS-PROPERTY-LIST isn't safe, even with "safe accessors" like
GETF.  This bug has been acknowledged by Franz.

SLYNK-ALLEGRO:MAILBOX creates and stores mailboxes in this plist in a
lazy get-existing-or-create fashion.  Access is serialized with a
lock, but this can still lead two threads calling
SLYNK-ALLEGRO:MAILBOX to create two different mailboxes, intead of the
desired behaviour where one just getting the one create by the other.
This is because there's no guarantee all threads that use this plist
would use ithe lock.

The consequence is, sometimes, (1) a thread will send messages to a
mailbox where nobody is waiting on and (2) another thread is waiting
(with SLYNK-BACKEND:RECEIVE-IF) on a mailbox noone will ever send a
message to.

The fix is to use a weak-keyed hash-table *MAILBOXES* mapping threads
to mailboxes.  The variable is protected by a global lock,
*GLOBAL-MAILBOX-HT-LOCK*.  Since noone else uses the hash-table, no
extra mailboxes are created.

Additionally, we also avoid the use of MP:PROCESS-PROPERTY-LIST (in
SLYNK-ALLEGRO:THREAD-ID) for Allegro version that have
MP:PROCESS-SEQUENCE.  That seems a safer way to compute an unique
fixnum for each thread.

* slynk/backend/allegro.lisp (*global-mailbox-ht-lock*)
(*mailboxes*): New variables.
(mailbox): No longer use MP:PROCESS-PROPERTY-LIST, use *MAILBOXES*.
(thread-id, find-thread): Use
MP:PROCESS-SEQUENCE, when available.
(*process-plist-lock*): Rename from *ID-LOCK*.