Handle `readtable-case` :invert correctly
With `readtable-case` :invert, `tokenize-symbol-thoroughly` was blindly
inverting every character in a string, but :invert only inverts the case
when the entire token is the same case. So `SYMBOL` is read as `symbol`,
and vice versa, but `Symbol` is read as-is. This is now handled
correctly.

* slynk/slynk.lisp (casify-char): Remove.
(char-casifier): New function.
Takes a string and returns a function which will apply the appropriate
case to the characters in the string, according to `readtable-case`.
(tokenize-symbol-thoroughly): Use `char-casifier`.