Handle :INVERT readtables correctly when parsing symbols
Co-authored-by: João Távora <joaotavora@gmail.com>
Reviewed-by: João Távora <joaotavora@gmail.com>

When parsing symbols, the :INVERT clause in READTABLE-CASE as used in
TOKENIZE-SYMBOL-THOROUGHLY was blindly inverting every character in a
string.  However :INVERT only inverts the case when the entire token
is the same case.  So `SYMBOL` is read as `symbol`, and vice versa,
but `Symbol` is read as-is.  This is now handled correctly.

* slynk/slynk.lisp (casify-char): Remove.
(char-casifier): New function.
Takes a string and returns a function which will apply the appropriate
case to the characters in the string, according to `readtable-case`.
(tokenize-symbol-thoroughly): Use `char-casifier`.