Don't count of Emacs's make-text-button's side-effects
Up until recently, make-text-button would change the (possibly
literal) string passed as an argument by attaching text properties to
this.  SLY relied on this behaviour since it was the only way to have
its buttons render correctly on Emacs <= 24.4.

make-text-button is now purely functional, it does not destruct the
arguments given to it, and so SLY must be updated.

Read more in this post and in the surrounding messages
https://lists.gnu.org/archive/html/emacs-devel/2020-06/msg00154.html

This commit probably breaks compatibility to Emacs 24.

* sly.el (sly--compilation-note-group-button)
(sly-apropos-insert-symbol)
(sly-xref-button)
(sly-db-frame-button)
(sly-db-local-variable-button)
(sly-inspector-part-button): Don't count on make-text-button's
side effects.

* contrib/sly-mrepl.el (sly-mrepl--make-result-button):
(sly-mrepl--make-result-button): Don't count on make-text-button's
side-effects.

* lib/sly-buttons.el (sly-make-action-button): Don't count on side
make-text-button's side-effects.