Unbreak indentation test "indent-13"
This required fixing a thinko by Zach Shaftel <zshaftel@gmail.com>.

The test was _also_ breaking because it relied on a actual bug that
Zach fixed: "thenret" was mispelled "threnret" in
sly--lisp-indent-if*-keyword.  So I updated the test.  This is the
correct indentation of the dummy if* form

(if* (eq t nil)
   then ()
        ()
 elseif (dsf)
thenret x
   else (balbkj)
        (sdf))

* (sly--lisp-indent-if*-1): Fix thinko.

* test/sly-cl-indent-test.txt (indent-13): Fix bug.