Add user-facing variable for customization of symbol completion methods
* lib/sly-completion.el

(sly-symbol-completing-read-function, sly-read-symbol-name):
Add new var referenced in sly-read-symbol-name for customization of
symbol completion methods. This allows users to, say, integrate ivy's
dynamic completions support into sly. Default value is
\#'sly-default-symbol-completing-read.

(sly-default-symbol-completing-read):
Encapsulate the logic for completing-read with the special sly
minibuffer in a function. This is to provide a default value for the
added configuration variable, and allow users to reset it.