Fix #313, #319, #335: Unbreak REPL reading
When asked to CL:READ very often, the REPL will transition quickly
between :read and :finished-reading modes of the :set-read-mode
channel method.  To give a change for input to come in before blocking
if off during the read, the function sly-mrepl--accept-process-output
was called at the top of the channel method.  But because that call
would do accept-process-output, it would not only let input in, but
sometimes also let the "next" invocation of the channel method happen
recursively, confusing the state of the important sly-mrepl--read-mark
variable, leading to a hang.

* contrib/sly-mrepl.el (define-channel-method :set-read-mode): Rework.