Record arg printouts at right time in the SLY Trace Dialog
Previously the printed representations of objects that were displayed
in the SLY Trace Dialog (which are sent from the Slynk side to Emacs
when the trace is collected) were taken at the time of the trace
collection, not at the time of the trace recording, which makes more
sense.

If an object mutates after its trace has been recorded, it wouldn't be
easy to detect that from the trace dialog.  This has now been fixed.

* contrib/slynk-trace-dialog.lisp (trace-entry): Add slots
printed-args and printed-retlist.
(initialize-instance): Record printed args.
(describe-trace-for-emacs): Use printed-args
(dialog-trace): Record a print of the return value at right time.