Fix regexps and variable naming in sly-cl-indent
Lots of changes here. The biggest is with respect to the LOOP keyword
regexps. The main reason I felt compelled to do this was that those
regexps weren't terminated with "\\_>" (symbol end), so `thereis` was
being indented like `the`. Then I just started fixing up a bunch of
other regexp related stuff while I was at it.

* lib/sly-cl-indent.el
(sly--lisp-loop-keyword-regexp): New function. Creates a regexp matching
a keyword or (maybe interned) symbol whose name is exactly one of its
arguments. It's now used to generate the values of the following variables.
(sly--common-lisp-body-introducing-loop-macro-keyword): Renamed to
`sly--lisp-indent-body-loop-macro-keyword`.
(sly--common-lisp-accumlation-loop-macro-keyword): Renamed to
`sly--lisp-indent-accumulation-loop-macro-keyword`.
(sly--common-lisp-prefix-loop-macro-keyword): Renamed to
`sly--lisp-indent-prefix-loop-macro-keyword`.
(sly--common-lisp-indent-clause-joining-loop-macro-keyword): renamed to
`sly--lisp-indent-joining-loop-macro-keyword`.
(sly--common-lisp-indent-indented-loop-macro-keyword): Renamed to
`sly--lisp-indent-indented-loop-macro-keyword`, and added some missing
keywords (probably still missing a few).
(sly--common-lisp-indenting-loop-macro-keyword): Renamed to
`sly--lisp-indent-conditional-loop-macro-keyword`.
(sly--lisp-indent-loop-macro-else-keyword): Renamed to
`sly--lisp-indent-else-loop-macro-keyword`.
(sly--lisp-indent-lambda-list-keywords-regexp): Use `regexp-opt`, cuz
why not.
(sly--lisp-indent-feature-expr-regexp): Fixed the regexp (#!+ is not a
valid feature expression as far as I know).
(sly--lisp-indent-loop-advance-past-keyword-on-line): Use
`skip-syntax-forward`, instead of `looking-at` every character in a
loop.
(sly--lisp-indent-function-lambda-hack): Fixed the regexp. Before it
would treat `(function-foo (lambda ()))` the same as
`(function (lambda ()))`.
(sly--lisp-indent-beginning-of-defmethod-qualifiers): Trivial change to
the regexp, and use `match-beginning` instead of `match-string` to test
which group matched.