Fix #385: Ensure that autodoc doesn't make stale continuations
After having learned that the eldoc-documentation-function runs within
while-no-input, it becomes clear that this is a hazard for anything
that calls sly-async.  So we must be careful to re-enable inhibit-quit
in sly-autodoc.

Maybe we should even enable inhibit-quit in sly-dispatch-event, for
atomicity, but for now do the safest thing that fixes the bug.

* contrib/sly-autodoc.el (sly-autodoc): Set inhibit-quit to t.

* sly.el (sly-dispatch-event): Add comment.