Fix #394: Revert "Don't force mREPL to use channel thread for everything"
This reverts commit ae4bbcaadc2b65d4f64c99b956c739372eeba1cc, which
was a total thinko.

Although what was written in that commit message is true for most
requests -- that :emacs-rex requests issued from the REPL can normally
be spawned in a new worker thread -- it (1) isn't true of some
slynk-mrepl SLYFUNs which check if the worker thread really is the
thread of the object representing the REPL, and (2) it most certainly
isn't true of the :emacs-interupt message.

The restriction of (1) is probably artificial, since these SLYFUNs
(which modify REPL history, for example) are "seeing" local REPL
bindings anyway via WITH-LISTENER-BINDINGS, and it probably wouldn't
hurt a ton to lift it.  But the restriction of (2) is imperative: the
purpose of that message is to interrupt a specific thread.

In the end, that commit was done to get a better experience in Clasp,
who seems to have a hard problem hanging on to thread IDs because of a
garbage collecting issue.  Let's rather have that fixed in Clasp.