SLIME deactivation: Only when really activated/loaded, prompt once or none
Deactivating SLIME was always done if `slime-lisp-mode-hook' was not in
`lisp-mode-hook' or the process was not interactive, and it was done by calling
`slime-mode'... which triggers automatic loading of SLIME, defeating the
already imperfect deactivation (which, for some SLIME's contrib, can't be done
easily).

Fix this by removing `slime-lisp-mode-hook' only if present on the hook, and
deactivating `slime-mode' in buffers only if really activated
(which means SLIME has been loaded).

Prompt the user only once for all occurences, and none if there are none.