Per #485: Don't use SLYNK package's symbols in ABCL backend
Using symbols from the SLYNK package in the ABCL backend fails because
the SLYNK package hasn't been created yet at the time the backend file
is CL:READ.

Though Backends may use the SLYNK-BACKEND and some other utility
packages, they should not rely on the presence of the SLYNK package,
certainly not at read-time.

Nevertheless, some -- like abcl.lisp -- qend up doing so anyway.  This
is probably because they were developed interactively in an
environment where some version of SLYNK is already loaded beforehand
somehow (perhaps it's burned in the image).

This commit doesn't solve the underlying problem of replacing those
uses with something else, but it uses a common trick to allow the
abcl.lisp file to be CL:READ even before the SLYNK package is created.

It is presumed that, by the time the calls to SLYNK symbols are
useful, the SLYNK package is effectively already loaded.  The hack is
localized and aptly named EVIL-HACK so to identifies the culprits.

It makes using a different solution in the future easier, and
hopefully makes it clearer what's going on.

* slynk/backend/abcl.lisp: Rework.