Move handling of SBCL's package variance warning to an around-compile hook
Recompiling slynk-backend.lisp and slynk.lisp causes SBCL to emit package
variance warnings. Currently, these are handled by wrapping `(asdf:load-system
"slynk")`. Unfortunately, this does not cover use cases where slynk is loaded
as a dependency of another library. At least
[40ants-doc](https://github.com/40ants/doc) has added slynk as a dependncy,
making it much more likely that slynk will be recompiled outside of a call to
load slynk directly.

We can handle this by moving the logic to handle the package variance warning
to the around compile hook of the slynk-backend and slynk
files. :AROUND-COMPILE with a LAMBDA was introduced in [ASDF 2.019](https://gitlab.common-lisp.net/asdf/asdf/-/blob/9dc59bb27c5251e954d6847cc1b507c5f399da7b/doc/Changelog#L984).

If the minimum required ASDF version is bumped to something closer to ASDF 3,
more (better) options become available to handle this, as noted in the
comments.

The only observable change should be that prior to this commit,
`(asdf:load-system "slynk")` would signal a `uiop::compile-warned-warning`
condition, but after this commit no warning at all is signaled and nothing is
printed that indicates there was a warning. This is a bit unfortunate, as there
could be other, non package-variance, warnings that would be silently
missed. But it's the best I could come up with without requiring ASDF 3 or
using a symbol from SB-INT.