Fix #514: Unconfuse sly-package-fu--create-new-import-from
This function was trying to go again to the relevant DEFPACKAGE
definition, except that point IS already at such a place.  Apparently,
sly-current-package returns nil in some situations and so this
needless repetition was actually harmful.

Charles Jackson <charles.b.jackson@protonmail.com>, discovered this
problem and put forth a fix.  This is a simplified version of that
fix.

* contrib/sly-package-fu.el
(sly-package-fu--create-new-import-from): Don't go to source definition again.
(sly-package-fu--add-or-update-import-from-form): Clean up comments.
(sly-package-fu--search-import-from): Remove spurious newline.