Use WITH-STANDARD-IO-SYNTAX for reading symbol names
SLY commands stopped working when using named-readtables and :PRESERVE
case. Internal parts of SLYNK would start looking for symbols with
lowercase symbol-names, which is nonsensical.

Add robustness to symbol lookups of the form:

    (read-from-string "a-package::a-symbol")

by giving them stable reader settings:

    (with-standard-io-syntax ...)

so that they work the same independent of the reader settings in the
image.

* slynk/backend/abcl.lisp: Use WITH-STANDARD-IO-SYNTAX.
* slynk/backend/allegro.lisp: Use WITH-STANDARD-IO-SYNTAX.
* slynk/backend/clasp.lisp: Use WITH-STANDARD-IO-SYNTAX.
* slynk/backend/lispworks.lisp: Use WITH-STANDARD-IO-SYNTAX.
* slynk/backend/mkcl.lisp: Use WITH-STANDARD-IO-SYNTAX.
* slynk/slynk-apropos.lisp: Use WITH-STANDARD-IO-SYNTAX.
* slynk/slynk-backend.lisp: Use WITH-STANDARD-IO-SYNTAX.
* slynk/slynk-source-path-parser.lisp: Use WITH-STANDARD-IO-SYNTAX.
* slynk/slynk.asd: Use WITH-STANDARD-IO-SYNTAX.

Co-authored-by: João Távora <joaotavora@gmail.com>