Per #564: Add SLYNK-BACKEND:FIND-SYMBOL2 for ad-hoc symbol-finding needs
Slynk backends use a decent amount of ad-hoc symbol finding.  This is
due to some backend files, which are READ before slynk.lisp (and the
SLYNK package is defined is loaded occasionally need to check things
in the Slynk packge (at run-time, of course).

This is a common source of problem, so its a good idea to do all this
symbol-finding behind an abstraction, which is the new
SLYNK-BACKEND:FIND-SYMBOL2.  Currently the implementation of that
sucks, but maybe that can be fixed later.

* slynk/slynk-backend.lisp: New SLYNK-BACKEND:FIND-SYMBOL2.
* slynk/backend/abcl.lisp: Use SLYNK-BACKEND:FIND-SYMBOL2.
* slynk/backend/allegro.lisp: Use SLYNK-BACKEND:FIND-SYMBOL2.
* slynk/backend/clasp.lisp: Use SLYNK-BACKEND:FIND-SYMBOL2.
* slynk/backend/lispworks.lisp: Use SLYNK-BACKEND:FIND-SYMBOL2.
* slynk/backend/mkcl.lisp: Use SLYNK-BACKEND:FIND-SYMBOL2.
* slynk/slynk-apropos.lisp: Use SLYNK-BACKEND:FIND-SYMBOL2.
* slynk/slynk-source-path-parser.lisp: Use SLYNK-BACKEND:FIND-SYMBOL2.