Fix #581: Reset output stream column when flushing listener streams
Close #583.

Streams, whether gray or not, have a so-called output-column or
line-column.  This change makes sure to reset it to 0 when the stream
is flushed, fixing the REPL problems described in #581.

In slynk/slynk-gray.lisp a new generic RESET-STREAM-LINE-COLUMN is
added, with a method for SLY's gray stream couple class,
SLY-OUTPUT-STREAM, and a couple of extra methods for SBCL and CMUCL,
which aren't always using gray streams in the context where
stream-flushing is needed.

This allows us to call RESET-STREAM-LINE-COLUMN safely in
slynk/slynk.lisp's FLUSH-LISTENER-STREAMS (except for abcl, which has
some problem I didn't investigate).

* slynk/slynk-gray.lisp (reset-stream-line-column): New generic and
  methods.

* slynk/slynk.lisp (flush-listener-streams): Call
  reset-stream-line-column.

Commit-message-by: João Távora <joaotavora@gmail.com>