Fix #595: Unbreak clisp which doesn't use :spawn comm style
The default value for SLYNK-MREPL::*DEDICATED-OUTPUT-STREAM-BUFFERING* depends on
the comm style and, previously, a rare but possible 'nil' comm style (as in Clisp)
would also prevent a gray stream from being used for the dedicated output stream,

So a native stream would be used instead.  This is possible, but only
on sbcl and cmucl.  So always make a gray stream, regardless of the
value of SLYNK-MREPL::*DEDICATED-OUTPUT-STREAM-BUFFERING*.

Fix #595, Fix #600.

* contrib/slynk-mrepl.lisp (open-dedicated-output-stream): Always
use line-buffering for the dedicated output stream.