sly.el(sly-eval-print): Apply the comment-region customizations.
If there's no prefix arg (1) in the call to comment-region, then
user-specified comment function does whatever the user customized it
to do.

And then, SLY indents single-colon comments weirdly so it's not
optimal to use exactly one colon for long code-result listings—it
needs too much additional work to indent/rewrite it properly.