Fix #681: Fix try-completion call
Can't just return the value of try-completion when using anything but
the prefix style because that will just fail even though there are
lots of possible completion when using the flex style.  Do what the
code did before and return the pattern.

This fix still maintains the intended behaviour for #636, which was
fixed by:

    commit 0ebe705c08af7126e0b2ce66a33fc7d23864b193
    Author: João Távora <joaotavora@gmail.com>
    Date:   Sat Apr 13 16:45:24 2024 -0500

        Fix #636: Behave more like vanilla with sly-simple-completions

* lib/sly-completion.el (sly--completion-function-wrapper): Tweak.